requirements.txt
flask
flask-sqlalchemy
psutil
pymysql
langchain
transformers
torch